backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-multipart==0.0.12
python-dotenv==1.0.1
slowapi==0.1.9
aiofiles==24.1.0
python-magic==0.4.27
Pillow==11.0.0
email-validator==2.2.0
# Email service
aiosmtplib==3.0.2
jinja2==3.1.4
# Device detection
user-agents==2.2.0
# Enhanced security
itsdangerous==2.2.0
